backend/requirements.txt
# NeuroApp – complete dependency list
# Core audio processing
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
praat-parselmouth==0.4.3

# Speech recognition & transcription
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Emotion recognition
transformers==4.36.2
datasets==2.16.1

# Audio feature extraction
python-speech-features==0.6
webrtcvad==2.0.10

# Data handling
pandas==2.1.4
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# API & Web
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.9

# OpenAI (flashcards / action item extraction)
openai==1.12.0

# Async HTTP (for testing)
httpx==0.26.0
pytest-asyncio==0.23.3

# Database (for future phases)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# OCR
pytesseract==0.3.10
Pillow==10.2.0
